Movement Experiments
Created two scenes:
     Movement_Cube
     Movement_Fish

Both use the same movement script to translate the object one direction across the game world. Fish scene uses some assets from another project of mine that came with a basic swimming animation so we could see the process of adding an animation with movement. In both scenes the object will move through the center cube, creating the scenario for Anthony to check for collision