Add files via upload
This version contains a large portion of the Turlington building geometry and an example of what photo-matching will look like. 
Before flooding git with a bunch of image files, this model should be tested for compatibility because Unity apparently does not like to import lines and some other sketchup features.