Camera and Player Controls Fixes
Made a change to Autocontrol to make the camera reset to the default relative position upon deactivating camera controls.

Switched camera rotation controls to be more intuitive ("R" moves the camera left, "Y" moves the camera right)

Deactivated unused HUD components (which were for some reason re-activated)

Commented-out GamemodeText functions in HUDController.cs since we are not using the GamemodeText component of the HUD.

Made ThirdPersionUserControl_AlexVersion.cs, which contains a fix to make the controller compatible with the new camera controls. Previously, the game controller moved the player in the direction of the camera's forward vector, resulting in the player running in a circle endlessly. Now, the game controller will move the player in the direction of the player's forward vector. I updated the Player prefab to use this version of the script. The old script is still in the project directory, however.